rename mappers to swap (#1546)
