+++ qiskit/transpiler/passes/__init__.py
@@ -13,7 +13,7 @@ from .optimize_1q_gates import Optimize1qGates
 from .decompose import Decompose
 from .barrier_before_final_measurements import BarrierBeforeFinalMeasurements
 from .mapping.check_map import CheckMap
-from .mapping.basic_mapper import BasicMapper
 from .mapping.cx_direction import CXDirection
 from .mapping.unroller import Unroller
-from .mapping.lookahead_mapper import LookaheadMapper
+from .mapping.basic_swap import BasicSwap
+from .mapping.lookahead_swap import LookaheadSwap
